LPS-165021 Rename and move ImportMapsShimUtil
I move it to kernel.frontend package as there are other similar utils
there.

Also, I added a "Frontend" prefix to the class name as other Utils do.
