+++ portal-kernel/src/com/liferay/portal/kernel/frontend/esm/packageinfo
